added/updated/fixed some problems that were found, other problems are still present

+++ Bonus_Examples/Modified Library For Images/DFRobot_LED_Matrix_invaders1/DFRobot_HT1632C.cpp
@@ -461,7 +461,7 @@ void DFRobot_HT1632C::drawImage(const byte * img, uint8_t width_t, uint8_t heigh
       }
       
       uint8_t copyInNextStep = 8 - max((src_y & 0b111), (dst_y & 0b111));
-      copyInNextStep = min(copyInNextStep, (height - src_y));
+      copyInNextStep = min(copyInNextStep,(uint8_t)(height - src_y));
       uint8_t dst_copyMask = (0b1 << copyInNextStep) - 1;
       dst_copyMask <<= (8 - (dst_y & 0b111) - copyInNextStep);
       uint8_t copyData = pgm_read_byte(&img[img_offset + (bytesPerColumn * src_x) + (src_y >> 3)]) << (src_y & 0b111);
@@ -503,7 +503,7 @@ void DFRobot_HT1632C::drawImageStr(const byte * img, uint8_t width_t, uint8_t he
       }
       
       uint8_t copyInNextStep = 8 - max((src_y & 0b111), (dst_y & 0b111));
-      copyInNextStep = min(copyInNextStep, (height - src_y));
+      copyInNextStep = min(copyInNextStep,(uint8_t)(height - src_y));
       uint8_t dst_copyMask = (0b1 << copyInNextStep) - 1;
       dst_copyMask <<= (8 - (dst_y & 0b111) - copyInNextStep);
       uint8_t copyData = pgm_read_byte(&img[img_offset + (bytesPerColumn * src_x) + (src_y >> 3)]) << (src_y & 0b111);
@@ -853,9 +853,3 @@ void DFRobot_HT1632C::doLength(const char text[]){
     ++i;
   }
 }
-
-
-
-
-
-
